+++ src/lib/components/table/public-api.ts
@@ -1,10 +1,12 @@
 export { TableComponent as BeyTableComponent } from './table.component';
 export {
+    BadgeTableCell as BeyBadgeTableCell,
     CellType as BeyCellType,
     LinkTableCell as BeyLinkTableCell,
     TableCell as BeyTableCell,
     TextTableCell as BeyTextTableCell
 } from './models/table-cell.model';
+export type { TableBadge as BeyTableBadge } from './models/table-cell.model';
 export {
     TableColumn as BeyTableColumn,
     TableConfig as BeyTableConfig,